Focus first invalid element during rejected submission
Re-introduce the built-in behavior provided by browser-level validation.

When a field is determined to be invalid during a client-side submission
validation, focus the first field that is invalid. When multiple fields
are invalid, do not focus fields after the first.

Similarly, when validating on `blur` events, **do not** focus, since
focus is moving manually on behalf of the user.