Fix HTML attribute bug when building `<select>`
While the `@options` instance variable generally represents the HTML
attributes for a given form field, building `<select>` elements poses an
exception to that rule. Since the final two arguments for [select][]
calls are `options = {}, html_options = {}`, the presence of
`@html_options` needs to be special-cased.

To account for that, change the monkey patches to utilize an
`attributes` local variable that references `@options` when
`@html_options` isn't available.

[select]: https://edgeapi.rubyonrails.org/classes/ActionView/Helpers/FormOptionsHelper.html#method-i-select