Improve support for custom Active Model validations
Prior to this commit, rendering a `<form>` with server-generated custom
Active Model validations would assign the message as
`Element.validationMessage`, but would _clear_ it from the HTML content.
This could be confusing when paired with `validateOn: ["blur",
"input"]`, because focus would be moved to the first invalid element
(which is not always the element with the custom validation message),
which could trigger the `blur` event, which could trigger a
re-validation.

Since the Active Model validation is custom and entirely server-side,
the client-side form would have no way to enforce it locally, so it
wouldn't be incorporated into the validation check. If the rest of the
form were invalid, it wouldn't prevent the re-submission. That means the
subsequent response would start the cycle over.

This commit ensures that the message is rendered into the HTML (in
addition to being processed as part of the Constraint Validations API).