Call both `.focus()` and `.scrollIntoView()` for invalid fields
To ensure that invalid elements are visible in the viewport, pair calls
to `.focus()` and `.scrollIntoView()`.

This shouldn't be necessary. In practice, some quirks have surfaced in
iOS Safari. The form control receives focus, but the scroll position
does not change. In desktop Safari, focus moves scroll position.

When combined and invoked sequentially, behavior is consistent across
iOS and Desktop Safari.